feat: added a greeting component and pictureFrame component
Practicing making components and making JSX
return statements. React components are just javascript functions
that return JSX which is HTML. After making it,
we use ES6 importing/exporting to import it into the main.jsx
and call the components. Note: all components must be
capitalized to differentiate between a react component
and a regular HTML statement. Importing and exporting
default and named works the same way in these cases.